+++ src/SmoothLingua/NLU/IntentPrediction.cs
@@ -6,4 +6,8 @@ internal class IntentPrediction
 {
     [ColumnName("PredictedLabel")]
     public string PredictedLabel { get; set; } = default!;
+
+    [ColumnName("Score")]
+    [VectorType]
+    public float[] Score { get; set; } = [];
 }